# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "App Name",
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/clients/{clientID}/telephones": {
            "get": {
                "tags": [
                    "Client Telephones"
                ],
                "summary": "Get a listing of the Clients.",
                "description": "Get all Client Telephones",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "clientID",
                        "in": "path",
                        "description": "id of Client",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Telephone"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Client Telephones"
                ],
                "summary": "Store a newly created Client Telephone in storage",
                "description": "Store Telephone from Client",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "clientID",
                        "in": "path",
                        "description": "id of Client",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Telephone that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Telephone"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Telephone"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/clients/{clientID}/telephones/{telephoneID}": {
            "get": {
                "tags": [
                    "Client Telephones"
                ],
                "summary": "Display the specified Telephone from Client",
                "description": "Get Client Telephone",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "clientID",
                        "in": "path",
                        "description": "id of Client",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "telephoneID",
                        "in": "path",
                        "description": "id of Telephone",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Client"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Client Telephones"
                ],
                "summary": "Update the specified Client telephone in storage",
                "description": "Update Client",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "clientID",
                        "in": "path",
                        "description": "id of Client",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "telephoneID",
                        "in": "path",
                        "description": "id of Telephone",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Telephone that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Telephone"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Telephone"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Client Telephones"
                ],
                "summary": "Remove the specified Telephone from Client from storage",
                "description": "Delete Telephone from Client",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "clientID",
                        "in": "path",
                        "description": "id of Client",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "telephoneID",
                        "in": "path",
                        "description": "id of Telephone",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/clients/{clientID}/telephones/{telephoneID}/primary": {
            "put": {
                "tags": [
                    "Client Telephones"
                ],
                "summary": "Set Client Telephone as Primary",
                "description": "Set Client Telephone as Primary",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "clientID",
                        "in": "path",
                        "description": "id of Client",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "telephoneID",
                        "in": "path",
                        "description": "id of Telephone",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Telephone"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/clients/{clientID}/telephones/primary": {
            "get": {
                "tags": [
                    "Client Telephones"
                ],
                "summary": "Set Client Telephone as Primary",
                "description": "Set Client Telephone as Primary",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "clientID",
                        "in": "path",
                        "description": "id of Client",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Telephone"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Client": {
            "required": [
                "name",
                "prefix"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "description": "name",
                    "type": "string"
                },
                "prefix": {
                    "description": "prefix",
                    "type": "string"
                },
                "user_id": {
                    "description": "user_id",
                    "type": "integer",
                    "format": "int32"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
        "Telephone": {
            "required": [
                "number"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "number": {
                    "description": "number",
                    "type": "string"
                },
                "telephoneable_id": {
                    "description": "telephoneable_id",
                    "type": "integer",
                    "format": "int32"
                },
                "telephoneable_type": {
                    "description": "telephoneable_type",
                    "type": "string"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        }
    },
    "securityDefinitions": {
        "oauth2": {
            "type": "oauth2",
            "description": "OAuth2 Implicit Grant",
            "flow": "implicit",
            "authorizationUrl": "/oauth/authorize",
            "tokenUrl": "/oauth/token",
            "scopes": {
                "scope": "Login to our system"
            }
        }
    }
}